Add more flexibility to getting layers from the Dataset.
Make chunk size default to 100 and compression level to 5 when creating a BAG.
Make chunk size a minimum of 100 when set.
Make sure chunking is enabled if it is set to a non-zero value.
Enable compression if it is a value from 1-10 (ignore 0).
Add type aliases to VR C structures.
Create VRNode layer more clearly.
Catch case of compression being set, but chunking not enabled.
Remove redundant initialising HDF5 layers to 0.